+++ plugins/postcss-media-minmax/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Media MinMax
 
+### Unreleased (patch)
+
+- Updated `@csstools/css-tokenizer` to `2.1.1` (patch)
+
 ### 1.0.1 (April 10, 2023)
 
 - Improve the detection of math function in media queries.