+++ plugins/postcss-media-minmax/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changes to PostCSS Media MinMax
 
-### Unreleased (patch)
+### 1.0.2 (April 10, 2023)
+
+- Updated `@csstools/css-tokenizer` to `2.1.1` (patch)
+- Updated `@csstools/css-parser-algorithms` to `2.1.1` (patch)
+- Updated `@csstools/css-calc` to `1.1.1` (patch)
+- Updated `@csstools/media-query-list-parser` to `2.0.4` (patch)
+
+### 1.0.1 (April 10, 2023)
 
 - Improve the detection of math function in media queries.